+++ backend/app.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const app = express();
-const Post = require('./models/post');
 const mongoose = require('mongoose');
+const postsRoutes = require("./routes/posts");
 
 mongoose.connect('mongodb+srv://wilson:e8LbS5YH0wMcPgLw@nodejsangular-wuf0b.mongodb.net/node-angular?retryWrites=true&w=majority') // database name: node-angular
   .then(() => {
@@ -28,60 +28,7 @@ app.use((req, res, next) => {
     next();
 });
 
-app.post('/api/posts', (req, res, next) => {
-  const post = Post({
-    title: req.body.title,
-    content: req.body.content
-  });
-  // console.log(post);
-  post.save().then(postData => {
-    res.status(201).json({
-      message: "Post added successfully!",
-      postId: postData.id
-    });
-  });
-});
-
-app.get('/api/posts', (req, res, next) => {
-  Post.find().then(documents => {
-    res.status(200).json({
-      message: 'Posts fetched successfully',
-      posts: documents
-    });
-  });
-
-});
-
-app.delete('/api/posts/:id', (req, res, next) => {
-  console.log(req.params.id);
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post deleted successfully"});
-  });
-})
-
-// patch only update the query, while put can create a new one
-app.put("/api/posts/:id", (req, res, next) => {
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content
-  })
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Update successfully!"});
-  })
-});
-
-app.get("/api/posts/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({message: "not found!"})
-    }
-  })
-})
+app.use("/api/posts",postsRoutes);
 
 module.exports = app;
