+++ src/app/posts/list/post-list.component.css
@@ -6,3 +6,8 @@
 .info-text {
   text-align: center;
 }
+
+
+mat-spinner {
+  margin: auto;
+}
