+++ src/app/posts/list/post-list.component.html
@@ -1,4 +1,5 @@
-<mat-accordion *ngIf="posts.length>0">
+<mat-spinner *ngIf="isLoading"></mat-spinner>
+<mat-accordion *ngIf="posts.length>0 && !isLoading">
   <mat-expansion-panel *ngFor="let post of posts">
     <mat-expansion-panel-header>
       <mat-panel-title>
@@ -12,4 +13,4 @@
     </mat-action-row>
   </mat-expansion-panel>
 </mat-accordion>
-<p class='info-text mat-body-1' *ngIf="posts.length==0">No post yet.</p>
+<p class='info-text mat-body-1' *ngIf="posts.length==0 && !isLoading">No post yet.</p>
