+++ src/app/posts/post/post-create.component.ts
@@ -15,6 +15,7 @@ export class PostCreateComponent implements OnInit{
   private mode = "create";
   private postId: string;
   post: Post;
+  isLoading = false;
 
   constructor(public postService: PostsService, public route: ActivatedRoute) {};
 
@@ -23,7 +24,9 @@ export class PostCreateComponent implements OnInit{
       if (paramMap.has('postId')) {
         this.mode = 'edit';
         this.postId = paramMap.get('postId');
+        this.isLoading = true;
         this.postService.getPost(this.postId).subscribe(postData => {
+          this.isLoading = false;
           this.post = {id: postData._id, title: postData.title, content: postData.content};
         });
       } else {
@@ -37,6 +40,7 @@ export class PostCreateComponent implements OnInit{
     if (form.invalid) {
       return;
     }
+    this.isLoading = true;
     if (this.mode ==="create") {
       this.postService.addPost(form.value.title, form.value.content);
     } else {
