+++ src/app/posts/posts.service.ts
@@ -4,12 +4,13 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { stringify } from 'querystring';
+import { Router } from '@angular/router';
 
 @Injectable({providedIn: 'root'})
 export class PostsService {
   private posts: Post[] = [];
   private postsUpdated = new Subject<Post[]>();
-  constructor (private http: HttpClient) {
+  constructor (private http: HttpClient, private router: Router) {
   }
 
   getPosts() {
@@ -42,6 +43,8 @@ export class PostsService {
       post.id = id;
       this.posts.push(post);
       this.postsUpdated.next([...this.posts]);
+      this.router.navigate(["/"]);
+
     });
   }
 
@@ -67,6 +70,8 @@ export class PostsService {
       updatedPosts[oldPostIndex] = post;
       this.posts = updatedPosts;
       this.postsUpdated.next([...this.posts]);
+      this.router.navigate(["/"]);
+
     });
   }
